refactor: update testing utilities file name and import paths for consistency
