+++ src/app/app.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
-import { provideMockAuth0Service } from './shared/utils/testingUtils';
+import { provideMockAuth0Service } from './shared/utils/testing-utils';
 import { provideHttpClient } from '@angular/common/http';
 import { provideRouter } from '@angular/router';
