+++ src/app/pages/sbp/register/sbp-register.component.html
@@ -226,7 +226,7 @@
       <div class="mt-6 flex justify-center">
         <button
           type="submit"
-          class="rounded-md bg-blue-600 px-8 py-3 text-sm font-medium text-white shadow-sm hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
+          class="bg-biocommons-primary rounded-md px-8 py-3 text-sm font-medium text-white shadow-sm hover:bg-sky-900 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
         >
           Register
         </button>
