+++ src/app/pages/user/services/request-service/request-service.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RequestServiceComponent } from './request-service.component';
-import { provideMockAuth0Service } from '../../../../shared/utils/testingUtils';
+import { provideMockAuth0Service } from '../../../../shared/utils/testing-utils';
 import { provideHttpClient } from '@angular/common/http';
 import { provideRouter } from '@angular/router';
